Rename .core -> .models
