+++ fah_xchem/__init__.py
@@ -1,2 +1,2 @@
-from .core import *
 from .lib import *
+from .models import *
